pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datagest"
dynamic = ["version"]
description = "Factory data versioning desktop app"
requires-python = ">=3.11"
dependencies = [
  "PySide6>=6.6",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-mock>=3.0",
  "ruff>=0.6",
  "pyinstaller>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "version.APP_VERSION"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["src"]
